Masteragent_v1/Masteragent/src/requirements.txt
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
typing-extensions>=4.5.0
jsonschema>=4.17.0
pandas
pytest>=7.3.0
black>=23.3.0
mypy>=1.3.0
flake8>=6.0.0
sphinx>=6.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
typing-extensions>=4.5.0
jsonschema>=4.17.0
pytest>=7.3.0
black>=23.3.0
mypy>=1.3.0
flake8>=6.0.0
langgraph==0.0.15
streamlit
plotly
networkx
langchain-google-genai==0.0.5
pydantic==1.10.13
langchain==0.0.350
google-generativeai==0.3.1
python-dotenv==1.0.0


